Added support for MySQL migration
Summary:
This diff moves all database source classes to `source` folder.
Also, the `SchemaInfo` is moved into separate files so it
can be reused for other source implementations.
For MySql support new `MysqlClient` class is implemented
for connecting and querying the source database alongside
the `MysqlSource` which implements the same functionality
as `PostgresqlSource`.
This diff adds new dependency, `libmysqlcppconn`, which provides
C++ interface for communicating with the MySql databases.

Reviewers: mferencevic

Reviewed By: mferencevic

Differential Revision: https://phabricator.memgraph.io/D2813